core: shrink public surface, evict vestigial output-format types
Make @policystack/core simpler without touching the frozen 1.0 contract:

- index.ts: drop ~22 internal-only re-exports — the 15 document builder
  helpers (keep visit / AST_VERSION), shouldEmit, seedCompany,
  deriveConsentMechanism, resolveJurisdiction, consentModelFor,
  ISSUE_CODE_IDS. All verified zero external consumers; none re-exported
  by the SDK. They stay available to core internals via direct imports.
- documents/index.ts: drop the 15-builder re-export block; privacy.ts /
  cookie.ts already import from ./helpers directly, so the barrel is now
  a clean AST contract.
- helpers.ts: delete the fully-dead italic() / ol() builders (+ orphaned
  ItalicNode import). The frozen "italic"/ordered-list AST node types and
  renderer handling are untouched.
- types.ts: delete OutputFormat / CompileOptions — a vestige of the
  removed compilePolicy() file-output model; core compiles to an AST and
  never formats output. Relocated to packages/renderers (their sole
  consumer), exported there so compilePolicy()'s signature stays nameable.

Test follow-through: the OutputFormat freeze-test moves with the type
(core types.test.ts deleted, guard recreated in renderers index.test.ts);
core index.test.ts now imports shouldEmit from ./emit.

Verified: vp run -r build (15/15), vp check (0 lint), vp run -r
check-types (11/11, 0 errors — no consumer relied on a removed export),
vp run -r test (core 463/463, renderers 29/29).
